Generic NetworkInteractorPackage
