--- Sources/JSONRPC/RPCRequestConvertible.swift
@@ -1,3 +1,0 @@
-public protocol RPCRequestConvertible {
-    func asRPCRequest() -> RPCRequest
-}
