+++ Sources/WalletConnectNetworking/NetworkRequest.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public protocol NetworkRequest {
+    var method: String { get }
+    var tag: String { get }
+}
