+++ Sources/WalletConnectNetworking/RequestSubscriptionPayload.swift
@@ -1,11 +1,11 @@
 import Foundation
 import JSONRPC
 
-public struct RequestSubscriptionPayload: Codable, Equatable {
+public struct RequestSubscriptionPayload<Request: Codable> {
     public let topic: String
-    public let request: RPCRequest
+    public let request: Request
 
-    public init(topic: String, request: RPCRequest) {
+    public init(topic: String, request: Request) {
         self.topic = topic
         self.request = request
     }
