+++ Sources/WalletConnectNetworking/ResponseSubscriptionPayload.swift
@@ -1,12 +1,14 @@
 import Foundation
 import JSONRPC
 
-public struct ResponseSubscriptionPayload: Codable, Equatable {
+public struct ResponseSubscriptionPayload<Request: Codable, Response: Codable> {
     public let topic: String
-    public let response: RPCResponse
+    public let request: Request
+    public let response: Response
 
-    public init(topic: String, response: RPCResponse) {
+    public init(topic: String, request: Request, response: Response) {
         self.topic = topic
+        self.request = request
         self.response = response
     }
 }
